Move code for retrieving buffer path into its own file
